refactor: split Artist Mix services
